requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
textstat==0.7.3
nltk==3.8.1
pandas==2.1.3
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
Pillow==10.1.0


email-validator==2.1.0
python-docx==1.1.0
pymupdf==1.23.6
pytest-asyncio==0.23.5
